Move typing.py -> _types.py
Avoid conflict with built-in typing module.
